requirements.txt
Django==3.2
djangorestframework==3.13.1
drf-yasg==1.21.3
pytest==7.1.2
pytest-django==4.5.2
requests==2.28.1
psycopg2-binary==2.9.3
